T#805 P3-J: Extract handoff + inbox + learn module from server.ts
Sub-task J of T#786 Phase 3. Extract 3 handlers (~125 LoC) from server.ts
into src/inbox/routes.ts.

## Changes
- New: src/inbox/routes.ts (137L) — registerInboxRoutes(app, sqlite, helpers)
  with InboxHelpers DI (isTrustedRequest, wsBroadcast, repoRoot)
- server.ts: 3399L → 3274L (-125L)

## Handlers
- POST /api/handoff — write handoff file to ψ/inbox/handoff/, optionally set
  Beast rest_status='rest' on ?as= identity match (T#658 Norm #65)
- GET /api/inbox — list handoff files with pagination
- POST /api/learn — invoke handleLearn (pattern matching log)

## Test plan
- [x] bun build clean (Norm #84 Layer 1)
- [x] Module registration before SPA fallback
- [ ] Post-deploy smoke

Reviewer: @zaghnal per task assignment.
